perf(): replaced the magnetic field strength data type from float to int so that the data displayed in the window does not overlap each other

+++ src/form.cpp
@@ -66,7 +66,7 @@ void Form::handlePressUnitChange(QString unit){
     ui->current_pressure_unit->setText(unit);
 }
 
-void Form::getMagneticField(QVector<float> magneticMg){
+void Form::getMagneticField(QVector<int> magneticMg){
     ui->magnetic_x->setText(QString::number(magneticMg[0]));
     ui->magnetic_y->setText(QString::number(magneticMg[1]));
     ui->magnetic_z->setText(QString::number(magneticMg[2]));
